gwang_feat: OngoingDashboard

+++ src/main/java/com/pigs/holiday/domain/Club.java
@@ -30,7 +30,7 @@ public class Club extends AuditingFields {
     int totalWins;
     int totalDraws;
     int totalLosses;
-    int mannerScore;
+    double mannerScore;
 
     @OneToMany(mappedBy = "homeClub")
     private List<MatchHistory> homeMatchHistoryList = new ArrayList<>();
@@ -51,7 +51,7 @@ public class Club extends AuditingFields {
     private List<MatchPost> awayMatchPostList = new ArrayList<>();
 
     protected Club(){}
-    private Club(String username, String password, String name, String university, String phone, String email, String clubName, String description, String region, String sportCategory, int totalMatches, int totalWins, int totalDraws, int totalLosses, int mannerScore) {
+    private Club(String username, String password, String name, String university, String phone, String email, String clubName, String description, String region, String sportCategory, int totalMatches, int totalWins, int totalDraws, int totalLosses, double mannerScore) {
         this.username = username;
         this.password = password;
         this.name = name;
@@ -68,7 +68,7 @@ public class Club extends AuditingFields {
         this.totalLosses = totalLosses;
         this.mannerScore = mannerScore;
     }
-    public static Club of(String username, String password, String name, String university, String phone, String email, String clubName, String description, String region, String sportCategory, int totalMatches, int totalWins, int totalDraws, int totalLosses, int mannerScore) {
+    public static Club of(String username, String password, String name, String university, String phone, String email, String clubName, String description, String region, String sportCategory, int totalMatches, int totalWins, int totalDraws, int totalLosses, double mannerScore) {
         return new Club(username, password,  name, university, phone, email, clubName, description, region, sportCategory, totalMatches, totalWins, totalDraws, totalLosses, mannerScore);
     }
